requirements.txt
python-telegram-bot==13.15
selenium==4.10.0
webdriver-manager==3.8.6
requests==2.31.0
pymongo==4.3.3
python-dotenv==1.0.0
schedule==1.2.0 